backend/requirements.txt
# LiveKit
livekit-agents>=1.0.0
livekit-plugins-deepgram>=1.0.0
livekit-plugins-openai>=1.0.0
livekit-plugins-cartesia>=1.0.0
livekit-plugins-silero>=1.0.0
livekit>=1.0.0

# HTTP client
httpx>=0.27.0
tenacity>=8.2.0

# Structured logging
structlog>=24.0.0

# Health check server
fastapi>=0.109.0
uvicorn>=0.27.0

# Environment
python-dotenv>=1.0.0
